requirements.txt
# Core dependencies
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.21.1,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# Google ADK and A2A
google-adk>=0.6.0,<1.0.0
google-generativeai>=0.5.0,<1.0.0

# Database
supabase>=1.0.3,<2.0.0
asyncpg>=0.27.0,<1.0.0
psycopg2-binary>=2.9.5 # For psql command in CI and potentially by Supabase lib

# Utils
httpx>=0.24.1,<1.0.0
tenacity>=8.2.2,<9.0.0
structlog>=23.1.0,<24.0.0

# Testing
pytest>=7.3.1,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
# Add other testing utilities like 'requests' for HTTP testing if needed
# For coverage:
# pytest-cov
